Save secrets to file and pass to Terragrunt
After fetching the secrets from KMS using Credstash we want to save them to a
file and pass that file as a -var-file to Terragrunt so it'll be included
in a Terraform run.

We'll be resolving and passing the file regardless content (even if it's
empty), so that teams don't have to enable it manually.

JIRA: PLAT-802