Reworks the exception hierarchy for the project
We had only a `UserError` as an exception parent, which was meant for
exceptions to be displayed to the user. It used a fixed string message
which wasn't always appropriate.

This introduces a `CDFlowError` at the top level and has `UserFacingError`
and `FixedMessageError` inheriting from that - `UserFacingFixedMessageError`
inherits from both of those:

```
                CDFlowError
                     |
UserFacingError             FixedMessageError
                \          /
          UserFacingFixedMessageError
```

JIRA: PLAT-848