Move state file to root after initialisation
The initialisation has to be done within the infrastructure directory
but the other commands can, and are, run from the root and the
directory is passed as an argument.

We also remove the quotes around the backend config parameters as
they are not removed when not running in the shell and so the
parameters aren't recognised.