Handle case when using us-east-1
We had a bug (or feature) in our code which prevented us from using
us-east-1 region, as it'd fail due to LocationConstraint not being
supported in us-east-1 (if we don't pass the constrain, by default
bucket is getting created there).  This commit fixed this allowing us to
use us-east-1 again.