Adding logging around the release of Lambda
When we run a release of a lambda project, we do a lot of work in the
background that isn't really exposed to the user so they get no feedback
when things go as expected. We've added logging to show whats going on
including some debug messages around the s3 bucket creation and tagging.

JIRA:PLAT-924