Make Deploy API reflect desired release metadata
We'd like to persist as much of the metadata about a release in the
release bundle and make that a valid tfvars file to pass straight to
Terraform. To that end we don't want to pass component, version etc.
but we do want to pass the path to the release metadata file.

JIRA: PLAT-945