Improve messaging
This change removes the name CDFlow Commands, as well as spelling
cdflow in lowercase (CDFlow looks a bit klunky to me). This is in
preparation for proposing we rename this repo to cdflow-docker-image to
more directly describe its purpose. It also includes more wording that
describes this repo's relationship with the cdflow repo.