Add support for tfstate-filename manifest key
When migrating older cdflow projects to new ones and trying to run the
new pipeline with new infrastructure definition in parallel, the deploy
finds the existing tfstate file and removes all of the infrastructure
(not very helpful). This change adds an optional `tfstate-filename` key
that chnages the last part of the key we use to store the tfstate file
in s3 so that the new infrastructure can coexist with the old, until we
are ready to move it to the default location.