Using lambda bucket name from account-scheme.json (#74)
We were trying to create buckets with the same name that we've already
created. This should solve that for each account. Now we will pull it
from the account-scheme.json. To ensure backwards compatability we set
the lambda bucket to empty if not in the account-scheme. This is to
deploy a lambda into the svcs account.

JIRA: PLAT-1325