Remove multiple account deploys feature
This feature was created before we could use organizations to get
account information. The idea was the account scheme would contain the
information required to switch role, which cdflow would pass to
terraform. The one service to use this is in the process of being
decommissioned, and future services will just use organizations to work
out accounts and roles to assume. Doing this now to make the code
simpler before adding support for release account.

JIRA: PLAT-1408