Persist account scheme through the pipeline
Store the account scheme in the release artefact and use this for
everything beyond finding the release artefact for deploys and destroys.
This will ensure changes we make to the account scheme will be applied
as a release goes through the pipeline, rather than (potentially) being
applied first to live.

JIRA: PLAT-1420