Support multi-region lambdas
In order to provision lambdas in multiple regions, the release zip needs
to uploaded to a bucket in each region. This change adds a
`multi-region` flag to the `cdflow.yml` manifest that causes the lambda
plugin to upload the zip to a set of buckets configured in the account
scheme. It then provides details of these buckets in the release
metadata passed to terraform.