Deploy account for secrets in new account schemes
We were going to put this in the release account, but since have decided
to deprecate credstash in favour of secrets manager in the deploy
account. This behaviour involves the minimum of work before we are able
to remove this code (in particular no need to migrate credstash secrets
when moving components to the new account schemes.